requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-compat==1.0.15
django-extensions==2.2.8
django-filter==2.2.0
django-hijack==2.1.10
django-hijack-admin==2.1.10
django-webpack-loader==0.7.0
djangorestframework==3.11.0
idna==2.9
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
